Moves Confirmation modal and styles file into own folder and updates imports
