+++ src/components/AddResourceModal/ConfirmationSuccessModal/ConfirmationSuccess.module.scss
@@ -1,0 +1,5 @@
+.modal {
+  max-width: auto;
+  margin: 0 auto;
+  padding: 0;
+}
