+++ src/components/Toolbar/Toolbar.js
@@ -28,7 +28,7 @@ import { ReactComponent as FoodIcon } from '../icons/CircleFoodIcon.svg';
 import { ReactComponent as ForagingIcon } from '../icons/CircleForagingIcon.svg';
 import { ReactComponent as WaterIcon } from '../icons/CircleWaterIcon.svg';
 import { ReactComponent as ContributeIcon } from '../icons/ContributeIcon.svg';
-import ConfirmationModal from '../AddResourceModal/ConfirmationModal.jsx';
+import ConfirmationModal from '../AddResourceModal/ConfirmationSuccessModal/ConfirmationModal';
 import { ReactComponent as FilterIcon } from '../icons/FilterIcon.svg';
 import { ReactComponent as ResourceIcon } from '../icons/ResourceIcon.svg';
 import { ReactComponent as SearchIcon } from '../icons/SearchIcon.svg';
